fix: change sale_price column type to NUMERIC(19,4)

+++ Backend/src/main/java/com/firomsa/inventory/v1/dto/SaleResponseDTO.java
@@ -1,5 +1,6 @@
 package com.firomsa.inventory.v1.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -13,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class SaleResponseDTO {
     private UUID id;
     private Integer quantity;
-    private Double salePrice;
+    private BigDecimal salePrice;
     private UUID productId;
     private String timestamp;
     private String message;
